feat(projects/art-jam-i): Refactors SVGDrawer to support styling from SVG file + overrides
Refactored SVGDrawer so that it picks up the styles directly from the SVG file,
but also added the ability to override with per-path overrides and global overrides.

The intent is that I want to be able to load a few more SVGs on top of the currently
loaded one, and that I want to be able to work on the colors of the SVG.

Also, renamed assets/profile-pic -> assets/portrait (more accurate name).