feat(projects/art-jam-i): Adds support for drawing SVG on WEBGL canvas via tr ansposition from an invisible 2D canvas as image
Rationale:
I want to add a video as an overlay on the canvas. This requires the use of a
shader (as per ChatGPT's direction), which requires the canvas to be in WEBGL mode.
Thus I needed to refactor the SVG to draw on the WebGl canvas.

Full attribution of GenAI-assistance available as part of this commit, next to
svg.js file.