feat(projects/art-jam-i): Scene 2 partial implementation
Migrates the rest of the original App.jsx to a second scene, scenes/scene-2.mjs.
With the new structure I can easily transfer existing state/render objects
across scenes and listen for events on the previous scenes to unlock
video playback.

First time the SVG layer renders on top of the video, yay!
No idea what makes the difference, but I think that having a clearer
code structure really helps to avoid issues due to lack of clarity in the prev
structure.

Now trying to play around with blending mode so that the SVG shows up
with contrast on the video.