feat+refactor(projects/mod-jam): Debugging + Refactor into split files
I've been trying to debug the physics-accurate tongue simulation code from ChatGPT,
so I've requested ChatGPT to provide a way to stop the simulation frame by frame
like I previously implemented for script.js.

Out of it came the Simulation class, which is used in the top-level draw function
to determine if the physics objects should be updated to the next 'frame' of the
simulation (next delta time). The simulation class is under src/physics/simulation.js

Additionally, in order to debug properly, I asked ChatGPT for a tracing utility
that can be used to make snapshots of the Frog body and tongue nodes, so that
a statistical analysis can be done on the frame-by-frame information. Out of it came
the Tracer class under src/utils/tracer.js.

That segways in the next change; the script2.js file was growing too large
(over 800 lines, with comments used to split the file sections). I've decided it
would be better for my sanity and speed of development to split the file into
a more conventional structure with a main file and a src directory.

I researched a bit about the support for ESM (modules) imports in the browser,
and modules are now supported by all browsers, so I opted for that to make sure
dependencies between files can be resolved properly from an entrypoint script
(the main file).