refactor(project/mod-jam): Integrates the FlyCounter back into the new project structure.
FlyCounter was the next logical choice to integrate after the SpeechSynthesizer,
simply because it's the only component that relies on it.

I chose to refactor the FlyCounter to a more generic Counter because I felt it would
be cleaner to put it under "ui/counter.js".
Also took the time to comment it better to understand why I decided to design the fly
counter as it is - a plain counter, with a voice over.

This refactor also required to integrate part of the previous keyPressed function
to be able to manually trigger the counter increment & listen to the easter egg
lines.