feat(projects/mod-jam): Adds a title screen
Adds a retro title screen with a blinking "Press {input} to Start"
I felt inspired to bring a sci-fi, old-school vibe to the title
because I'm already using an old NASA-sounding synthesizer voice,
and I think that fits well with 60s-80s space race theme.

The title image are made in photoshop - they are both simple transforms
of texts (perspective warp, blending options with gradient).
The inspiration for the title was the Super Metroid logo, and the
Star Wars intro text scroll.

In terms of implementation, what to say?
- Nothing in this commit is made with GenAI.
- src/utils/drawing.js:ConvenientImage is really just a run-of-the-mill
  OOP wrapping of p5.Image that makes loading an image just a little bit
  more dynamic and malleable.
- src/utils/drawing.js:blink does what it implies - it executes a draw
  function at intervals, for a given duration, to give the effect of blinking.
- src/globals.js was updated with better comments as a was doing a commenting sweep
  and adding a GLOBALS.SCENE variable.