feat(projects/mod-jam): Replaces hunger bar with oxygen reserve countdown
Decided to replace the HungerBar with this instead because I kept
wondering "What kills you in space?" and the same answer kept popping up:
asphyxiation.
I felt like that would be more appropriate than "dying of hunger", when
you are literally eating fly after fly as the core loop.
This feels more definitive, more accurate to the theme, and creates a sense
of inevitability and finality to it; no matter how many flies you eat,
you will die in this cold outer space.

Implementation details:

- Opted for using a font instead of displaying rects + triangles to save
  lots of time
- Opted for a monospace font to ease centering the countdown label
- Opted for a callback to call the synthesizer, keeps the concerns
  cleanly separated.