feat(projects/mod-jam): Draw the frog legs
Had to add the legs! Otherwise it's only a half frog. Now it's a proper frog.
Now it really feels like the frog is floating in space.
Also added a red dot on the center of mass of the frog so that I can check if
the rotation animation feels right. Not quite yet, but it's better than it was.
I'll have to touch this up if time allows.

My intention here is that if I get the time to do so, have the frog
shake its legs when the user adds torque to spin the frog faster. That would
be a great way to tell the user they are doing something.

If I wanted to maximize this idea, I'd also add that the frog can
extend or close-in its legs, so that it can benefit from a smaller inertia.
But that's really out of scope.