feat-wip(projects/mod-jam): Implements starry sky with zoom out
I reflected for a while on how I wanted the out-of-bounds behaviour to act
like. Did I want the frog to just go out of bound? Did I want it to wrap
around? Did I want to load new maps everytime the frog goes out of bound?

I came up to Dr Pippin Barr in class with the question, and we explored
the emotionality and player experience that each would provide, and
Pippin proposed a solution I had barely considered: Make the camera zoom
out; this would reinforce the terror of being alone, floating in space, because
no matter how far you go, you stay alone.

I really liked the idea, but more importantly, I think it really resonated with
me. In the end what make people want to engage with media is the contrast
in emotion that the media provide them. And I think designing with this in mind,
from the ground up, including mechanics, game design elements, etc, is what
can make memorable experiences. I really want to dig further in this approach
and better understand how to make emotionally and philosophically stimulating
games for my audiences.

---

Hence I implemented this little demo in starry-space.js that renders
stars in a pseudo-random, deterministic way (with a help from my little friend
ChatGPT), but that also zooms out as the player approaches the boundary of the
viewport.
Full documentation is available in the file.