feat(projects/mod-jam): Implements flies (draw, movement)
Added flies to the game - simple sprites looking like flies with a space suit.
I pondered what kind of movement would be interesting (and diegetic);
Given the nature of the simulation (being in space), I figured a smooth, linear movement
would fit better in context. I considered using Perlin noise, but I figured it would
not fit as well with the space context. Another possibility was to have the flies not
move at all (they can't use their wings in space), but that would go against the requirements.

Extracted the CoordinatesBox & CoordinatesPoint from the starry-sky.js file to allow the flies
to have bounding boxes that can be used to determine whether or not the fly is on screen.
This was used to avoid rendering flies when off screen.
While this has little implications on actual performance, I did this as a reflection on
how a real engine would handle drawing assets when out of camera viewport.

Flies have different movement (acceleration & how often they change direction) to give the
sense of the flies not having a cookie-cutter behaviour.