feat(projects/mod-jam): Adds Game Over Screen
Iterative implementation:
1. Added initial GameOverScreenOverlay which would only apply a tint filter
2. Modified all PhysicsObjectModel to support scaling the velocities down to add a slow-motion effect.
   I wasn't able to make it work, but the final outcome was still serendipitously satisfying.
3. Added a GAME OVER text + cause of death text + reset button + audio cue (GAME OVER utterance using text to speech api)

All committed at once because I wanted a somewhat self-contained commit with everything from the
Game Over feature.

---
Comments

This one is a bit of an incomplete feature; the tint filter does not produce the
expected outcome of having a sepia filter, but the gray filter is pretty fitting as well, so I left
as is.

The scale-down of velocity to add a slow-motion effect doesn't work either - the frog
just stops moving altogether, and the flies move weirdly. But the outcome is sufficiently
good to be acceptable, so I'll leave as is. I do suspect however that this may become an issue
later on when I implement the tongue. It may behave erratically once the scale-down is applied
(well, any more erratically than it already is).

The Game Over texts & button is really satisfying in my opinion, and while it took
2 hours to implement, I think it was a worthwhile investment.