docs(projects/variation-jam): Brainstormed ideas -> IDEAS.md
Brainstormed ideas for tag games.
Started from the obvious, and went into more meta versions of the game –
What does it mean to tag someone? What's the essence of the game?
Catching up to someone? Touching someone? Giving them something, 'it'?

The only thing I'm a bit fuzzy/scared about is how to implement the AI for
the characters, and also have difficulty settings for those AIs.