feat(projects/variation-jam): Updates styling of menu
Given the theme of the games, "tag", I thought it would be more on theme to use a font that
brings mischievousness, fun, and play to the table.
I thought of the Impossible Quiz, and decided to find a font that was similar to the impossible
quiz.
Following the same aesthetic, I switched the background back to white-ish, and used very basic
colors.