feat(projects/variation-jam): Modifies SceneManager to support multi-scene Games
Long story short, I wanted to have games with multiple scenes - because I want
an intro/instructions scene, a main gameplay loop scene, and a win/lose scene.

So what I did was to modify the SceneManager to support registering another SceneManager
as a child, and allow a child SceneManager to return a SceneRequest to a parent
SceneManager when it can't find the requested Scene. That way it provides an escape
hatch where a scene within a game can request to navigate back to the menu.

This is feeling more like a hacked-together SPA at this point lol
But I won't get any more complex than that (hopefully).
Now's the time to do games.