feat(projects/variation-jam): Adds centralized FontBook
As I was developing the first game, I realized that font loading is a cross-cutting concern,
and that fonts need not be loaded in every game.
So I centralized the font loading/caching mechanism in a class, FontBook.

That way all fonts can be loaded ahead of time, and games can request the font from
the FontBook; the FontBook provides sentinel values when font-loading has resulted in
an error, and as such, we don't call p5.textFont and fallback to the default font
without incurring an error.

I'm really surprised that p5.textFont doesn't support fallback and listing which fonts
are available. That seems like a pretty big DX miss.
Anyways. Now I have a robust way of detecting if the font loaded properly or not
and handling this cleanly if the font can't be found.