feat(projects/variation-jam): Clock Tag Game + Refactor typo styles
Implemented the Clock Tag Game as first game.
The Clock Tag Game is a simple game where you have 15 seconds for the seconds hand of the
clock to reach any other of the hands - thus being the tag, and tagging it.

I wanted to do the clock first, because it seemed like the simplest game to implement.
Pretty huge commit, but I did the clock on Tue 2025-11-18 night in p5 playground.
Bringing everything together with a humoristic approach - I added elevator music
for the game because it's essentially a waiting game, and added a Sonic the Hedgehog
sound bite if you lose to taunt the player into playing again.
Everything is really on-the-nose and wacky, in the spirit of Impossible Quiz.

I've also refactored the menu to use the typo styles from 'theme.js'
to standardize all the text size across the games. Took this opportunity to clean
it up a bit too.