feat(projects/variation-jam): Implemented NPC AI engine + Classic tag (partially) + Additional comments
In order to get the tag game to work, I had to implement an NPC AI engine - that is, a system
that given data, will automatically update the behaviour of the NPC so that the NPC reaches its objective.
Given this is a tag game, the main goal of the NPC is to pursue the player, or to evade the player.

This is my first steps into the world of in-game AIs and state machines.
I knew this would be a complex endeavour, and indeed it was!
This being said, in comparison to making the FrogFrogFrog physics-simulated tongue,
this was a lot easier.

Having learnt data structures in SOEN back in 2013, and knowing a good deal about Graphs and Graph
theory, I had a solid inkling that movement would be done using a distance calculation algorithm,
like Djikstra's distance algorithm, on a grid-based graph of nodes with edges to all their neighbours.

What was fuzzier was what would be the appropriate design patterns to employ to bridge this pathfinding
algorithm with the in-game NPC and the display logic.

The resulting engine was developed entirely using GenAI and can be found under `js/src/engine`.

---

Additionally, the Classic Tag game has been partially implemented - only to the extent that it integrates
the engine.

Finally, I've taken this massive commit as an opportunity to improve in-code documentation on all the files
affected by this commit.