feat(projects/variation-jam): Fixes NPC AI, integrates tag collision detection
Behaviours were updated through the use of GenAI.
Multiple passes were needed, ultimately settling for a "good enough" evasion
behaviour where grid nodes near the edges of the canvas are penalized.

Additionally, collision between the player and npc was added to the classic
tag play scene. I added an additional IdleBehaviour with key "idle" so that
the NPC stops for a second before starting chasing the player.
Furthermore, a sort of "mutex" was added to ensure that the player and
npc don't repeatedly exchange their mode (pursuer/evader) every frame.
Instead a cooldown of sort was implemented by passing a delayMs to
Player.setMode and NPC.setMode.

Player & NPC colors were also updated, because just switching between
pursuer and evader color was really confusing; the Player circle wouldn't
retain its visual identity when the modes were exchanged. So instead,
the player has a solid color, and the mode is communicated through
the stroke color.