feat(projects/variation-jam): Implements the Realistic Tag variation
The realistic tag is a joke variation that essentially tells the player
to go play tag in real life and come back to complete the game.

The implementation leverages the engine/input.js' MouseInput class
to capture mouse events, which it then uses to determine whether or not
instances of the components/button.js:Button class have been clicked.

The idea behind this variation was to cross beyond the boundary of the
virtual medium, a bit like an Alternate Reality Game
(ARG), like Cicada 3301.