feat(projects/variation-jam): Implements the zombie variation
The zombie variation was a lot easier to make, once the classic variation was
made. It reuses the same systems, with a few modifications to add additional
variance to the NPC behaviours. But it is essentially the classic-tag, with
many opponents, and additional mechanics for the mood.

Originally I wanted to make an actual schoolyard-like zombie game, where
it's just a tag, but I quickly realized that it would be too complex to implement
a system where the NPCs change target when they get close to a non-tagged NPC.
So instead I started by making all the NPCs zombies, and I realized... I liked
the effect it gave.
So from then on it was more about massaging the design & interaction to support
the horror-theme of being chased by zombies, and voila!
I've designed the zombies to be slow, and a bit stupid, but if you get too close,
they start rushing you very fast.
I've also turned the background to black, which clashes with the rest of the games,
and added a creepy atmosphere sound in the back.
Win/Lose scenes also have their own sound effects to support the mood.